Fix sinkShip method in gameBoardFactory
The findIndex method I was using returns -1 when an index for an element
is not found. This was leading to problems with the sinkShip method
where it would keep splicing the array even if the index was not found
anymore, leading to my game loop clearing out my ships array.